lambdas/geocoding/requirements.txt
psycopg2-binary==2.9.7
mapbox==0.18.1
